refactored components folder structure and updated file imports accordingly
