+++ src/Components/Navbar/Navbar.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import ThemeToggle from "./ThemeToggle";
+import ThemeToggle from "../ThemeBtn/ThemeBtn";
 
 type Props = {}
