+++ src/Components/StackSection/StackSection.tsx
@@ -1,8 +1,8 @@
 
-const TechStack = () => {
+const StackSection = () => {
     return (
         <section id="techstack" className="min-h-screen">TechStack</section>
     )
 }
 
-export default TechStack
+export default StackSection
