+++ src/app/page.tsx
@@ -1,20 +1,20 @@
-import AboutMe from "@/components/AboutMe";
+import AboutSection from "@/components/AboutSection/AboutSection";
 import ExperienceSection from "@/components/ExperienceSection/ExperienceSection";
-import Footer from "@/components/Footer";
-import Hero from "@/components/Hero";
-import Navbar from "@/components/Navbar";
-import Projects from "@/components/Projects";
-import TechStack from "@/components/TechStack";
+import Footer from "@/components/FooterSection/FooterSection";
+import Hero from "@/components/Hero/Hero";
+import Navbar from "@/components/Navbar/Navbar";
+import ProjectsSection from "@/components/ProjectsSection/ProjectsSection";
+import StackSection from "@/components/StackSection/StackSection";
 
 const Home = () => {
   return (
     <>
       <Navbar />
       <Hero />
-      <Projects />
-      <TechStack />
+      <ProjectsSection />
+      <StackSection />
       <ExperienceSection />
-      <AboutMe />
+      <AboutSection />
       <Footer />
     </>
   );
